requirements.txt
django
djangorestframework
django-filter
drf_yasg
PyYAML
coverage